refactor: fetcher 파일들 도메인 별 hook 파일 안으로 이동
- config는 원래대로 apis 바로 밑으로 이동
