--- frontend/src/apis/fetcher/members.ts
@@ -1,8 +1,0 @@
-import { API_PATH } from '../../constants/api';
-import { client } from '../config/axios';
-
-export const getMembersRequest = async () => {
-  const { data } = await client({ method: 'get', url: API_PATH.MEMBERS });
-
-  return data;
-};
