--- frontend/src/apis/fetcher/reservation.ts
@@ -1,12 +1,0 @@
-import { API_PATH } from '../../constants/api';
-import { client } from '../config/axios';
-
-export const createReservationRequest = ({ couponId, date, time }) =>
-  client({
-    method: 'post',
-    url: `${API_PATH.RESERVATIONS}`,
-    data: {
-      couponId,
-      startAt: `${date} ${time}:00`,
-    },
-  });
