+++ frontend/src/hooks/@queries/coupon.ts
@@ -1,5 +1,6 @@
 import { useMutation } from 'react-query';
-import { createCouponRequest } from '../../apis/fetcher/coupon';
+import { client } from '../../apis/axios';
+import { API_PATH } from '../../constants/api';
 
 export const useCreateCouponMutation = ({ receiverIds, content }, { onSuccess = () => {} } = {}) =>
   useMutation(() => createCouponRequest({ receiverIds, content }), {
@@ -7,3 +8,13 @@ export const useCreateCouponMutation = ({ receiverIds, content }, { onSuccess = 
       onSuccess();
     },
   });
+
+const createCouponRequest = async ({ receiverIds, content }) => {
+  const { data } = await client({
+    method: 'post',
+    url: API_PATH.SEND_COUPON,
+    data: { receiverIds, content },
+  });
+
+  return data;
+};
