+++ frontend/src/hooks/@queries/members.ts
@@ -1,6 +1,7 @@
 import { AxiosResponse } from 'axios';
 import { useQuery } from 'react-query';
-import { getMembersRequest } from '../../apis/fetcher/members';
+import { client } from '../../apis/axios';
+import { API_PATH } from '../../constants/api';
 import { UserProfile } from '../../types';
 
 const QUERY_KEY = {
@@ -9,3 +10,11 @@ const QUERY_KEY = {
 
 export const useGetMembers = () =>
   useQuery<AxiosResponse<UserProfile[]>>(QUERY_KEY.members, getMembersRequest);
+
+/** FETCHER */
+
+const getMembersRequest = async () => {
+  const { data } = await client({ method: 'get', url: API_PATH.MEMBERS });
+
+  return data;
+};
