+++ frontend/src/hooks/Main/queries/coupons.ts
@@ -1,5 +1,5 @@
 import { useQuery } from 'react-query';
-import { client } from '../../../apis/config/axios';
+import { client } from '../../../apis/axios';
 import { API_PATH } from '../../../constants/api';
 import { Coupon } from '../../../types';
