+++ frontend/src/index.tsx
@@ -8,11 +8,11 @@ import { css, Global } from '@emotion/react';
 import { QueryClientProvider } from 'react-query';
 import { ReactQueryDevtools } from 'react-query/devtools';
 import { RecoilRoot } from 'recoil';
-import { queryClient } from './apis/config/queryClient';
 import './assets/favicon/favicon.ico';
 import reset from './styles/GlobalReset';
 import global from './styles/GlobalStyled';
 import { ThemeProvider } from './styles/ThemeProvider';
+import { queryClient } from './apis/queryClient';
 
 const rootElement = document.getElementById('root')!;
 const root = createRoot(rootElement);
